+++ cmd/lgballtDiscordBot/main.go
@@ -11,7 +11,7 @@ import (
 	"time"
 )
 
-//go:generate python ../../scripts/prebuild.py ../.. ../../internal/buildInfo/buildInfo.go buildInfo 3.1.0
+//go:generate python ../../scripts/prebuild.py ../.. ../../internal/buildInfo/buildInfo.go buildInfo 3.2.0
 
 func main() {